requirements.txt
# AIRO v2.0 - Requirements
# Modernisierte Dependencies für AI-gestützte Code-Generierung (2025)

# === Core Dependencies ===
# Ollama Python Client für LLM-Interaktion
ollama>=0.4.0

# Environment Management
python-dotenv>=1.0.0

# === CLI & User Interface ===
# Rich für moderne, schöne CLI-Ausgabe
rich>=13.7.0

# === Code Quality Tools ===
# Ruff - Moderner, schneller Python Linter (ersetzt flake8, pylint, etc.)
ruff>=0.7.0

# MyPy für Type-Checking
mypy>=1.11.0

# Black für Code-Formatting
black>=24.8.0

# === Security Scanning ===
# Bandit für Python Security-Analyse
bandit>=1.7.9

# Safety für Dependency-Security-Check
safety>=3.2.0

# === Testing ===
# Pytest - Modernes Testing-Framework
pytest>=8.3.0
pytest-cov>=5.0.0  # Coverage-Reports
pytest-mock>=3.14.0  # Mocking

# === HTML Parsing (vom Original) ===
html5lib>=1.1

# === Optional: Advanced Features ===
# Für RAG-Funktionalität (optional)
# langchain>=0.3.0
# chromadb>=0.5.0

# Für erweiterte Code-Analyse
# astroid>=3.3.0

# Für Git-Integration
# gitpython>=3.1.43

# === Development Dependencies (Optional) ===
# ipython>=8.26.0  # Bessere REPL
# jupyter>=1.0.0   # Für Notebooks
